+++ dave/common/raw_container.py
@@ -11,26 +11,38 @@ from .sample_type import SampleType
 
 @dataclass
 class RawContainer:
+    """
+    A pickable representation of a container, sendable between the two dave process
+
+    This should contains every piece of information needed by the GUI
+    """
+
     id: int
     data: bytearray
     name: str
     original_shape: Tuple[int, int]
-    # dimensions are fixed when the container is 2D and both dimensions are therefore fixed
     dimensions_fixed: bool
     interleaved: bool
     sample_type: SampleType
-    # container_cls: Type
     default_layout: DataLayout
     possible_layout: List[DataLayout]
 
     @dataclass
     class InScopeUpdate:
+        """
+        Used to send a data update to the GUI
+        """
+
         id: int
         data: bytearray
         shape: Tuple[int, int]
 
     @dataclass
     class OutScopeUpdate:
+        """
+        Used to send an out-of-scope update to the GUI
+        """
+
         id: int
 
     def update(self, update: InScopeUpdate):